final-check-v1

+++ src/main/java/com/example/webmuasam/service/ProductService.java
@@ -1,21 +1,26 @@
 package com.example.webmuasam.service;
 
 import com.example.webmuasam.Specification.ProductSpecification;
+import com.example.webmuasam.dto.Response.ProductResponse;
 import com.example.webmuasam.dto.Response.ResultPaginationDTO;
-import com.example.webmuasam.entity.Category;
-import com.example.webmuasam.entity.Images;
-import com.example.webmuasam.entity.Product;
-import com.example.webmuasam.entity.ProductVariant;
+import com.example.webmuasam.entity.*;
 import com.example.webmuasam.exception.AppException;
 import com.example.webmuasam.repository.CategoryRepository;
 import com.example.webmuasam.repository.ImageRepository;
 import com.example.webmuasam.repository.ProductRepository;
 import com.example.webmuasam.repository.ProductVariantRepository;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Base64;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,28 +36,95 @@ public class ProductService {
         this.productVariantRepository = productVariantRepository;
         this.imageRepository = imageRepository;
     }
+    public ProductResponse convertProductToProductResponse(Product product) {
+        ProductResponse productResponse = new ProductResponse();
 
-    public Product createProduct(Product product) {
-        // Gán danh mục
+        productResponse.setId(product.getId());
+        long quantity = product.getVariants().stream().mapToLong(ProductVariant::getStockQuantity).sum();
+        productResponse.setQuantity(quantity);
+
+        productResponse.setName(product.getName());
+        productResponse.setDescription(product.getDescription());
+        productResponse.setPrice(product.getPrice());
+        productResponse.setCreatedBy(product.getCreatedBy());
+        productResponse.setUpdatedBy(product.getUpdatedBy());
+        productResponse.setCreatedAt(product.getCreatedAt());
+        productResponse.setUpdatedAt(product.getUpdatedAt());
         if(product.getCategories() != null) {
-            List<Long> categoryIds = product.getCategories().stream().map(x -> x.getId()).collect(Collectors.toList());
-            List<Category> categories = categoryRepository.findAllById(categoryIds);
-            product.setCategories(categories);
+            List<ProductResponse.CategoryPro> productResponseCategoryPro = product.getCategories().stream().map(c->{
+                ProductResponse.CategoryPro categoryPro = new ProductResponse.CategoryPro();
+                categoryPro.setId(c.getId());
+                categoryPro.setName(c.getName());
+                return categoryPro;
+            }).collect(Collectors.toList());
+
+            productResponse.setCategories(productResponseCategoryPro);
+        }
+        if(product.getImages() != null) {
+            List<String> base64Images = product.getImages().stream().map(img -> Base64.getEncoder().encodeToString(img.getBaseImage())).collect(Collectors.toList());
+            productResponse.setImages(base64Images);
+        }
+        Double totalStar = 0.0;
+        if (product.getReviews() != null && !product.getReviews().isEmpty()) {
+            totalStar = product.getReviews()
+                    .stream()
+                    .mapToInt(Review::getRating)
+                    .average()
+                    .orElse(0.0);
+            long quantityReview = product.getReviews().size();
+            productResponse.setQuantityReview(quantityReview);
+        }
+        productResponse.setTotalStar(totalStar);
+        if(product.getVariants() != null) {
+            List<ProductResponse.Variants> variants= product.getVariants().stream().map(c->{
+                ProductResponse.Variants variant = new ProductResponse.Variants();
+                variant.setId(c.getId());
+                variant.setColor(c.getColor());
+                variant.setSize(c.getSize());
+                variant.setStockQuantity(c.getStockQuantity());
+                return variant;
+            }).collect(Collectors.toList());
+            productResponse.setVariants(variants);
+        }
+        return productResponse;
+    }
+    @Transactional
+    public Product createProduct(String name, Double price, String description, List<MultipartFile> images,List<Long> categories) {
+        // Gán danh mục
+        Product product = new Product();
+        product.setName(name);
+        product.setDescription(description);
+        product.setPrice(price);
+        if(categories != null) {
+            List<Category> categorienews = categoryRepository.findAllById(categories);
+            product.setCategories(categorienews);
         }
 
 
         // Gán lại product cho images & variants
-        if (product.getImages() != null) {
-            for (Images image : product.getImages()) {
-                image.setProduct(product);
+        if (images != null) {
+            List<Images> imagesNew = new ArrayList<>();
+            for (MultipartFile image : images) {
+                try{
+                    Images img =new Images();
+                    img.setBaseImage(image.getBytes());
+                    img.setProduct(product);
+                    imagesNew.add(img);
+                }catch (IOException e){
+                    throw new RuntimeException("loi up anh",e);
+                }
             }
+            product.setImages(imagesNew);
         }
 
-        if (product.getVariants() != null) {
-            for (ProductVariant variant : product.getVariants()) {
-                variant.setProduct(product);
-            }
-        }
+//        if (productVariants != null) {
+//            List<ProductVariant> productNews = new ArrayList<>();
+//            for (ProductVariant variant : productVariants) {
+//                variant.setProduct(product);
+//                productNews.add(variant);
+//            }
+//            product.setVariants(productNews);
+//        }
 
         // Tổng quantity từ variants
         long totalQuantity = product.getVariants() != null
@@ -62,45 +134,46 @@ public class ProductService {
 
         return productRepository.save(product);
     }
-
-    public Product updateProduct(Product updatedProduct) throws AppException {
-        Product existingProduct = productRepository.findById(updatedProduct.getId())
+    @Transactional
+    public Product updateProduct(Long id,String name, Double price, String description, List<MultipartFile> images,List<Long> categories) throws AppException {
+        Product existingProduct = productRepository.findById(id)
                 .orElseThrow(() -> new AppException("Product not found"));
 
         // Cập nhật thông tin cơ bản
-        existingProduct.setName(updatedProduct.getName());
-        existingProduct.setPrice(updatedProduct.getPrice());
-        existingProduct.setDescription(updatedProduct.getDescription());
+        existingProduct.setName(name);
+        existingProduct.setPrice(price);
+        existingProduct.setDescription(description);
 
         // Cập nhật danh mục
-        if(updatedProduct.getCategories() != null) {
-            List<Long> categoryIds = updatedProduct.getCategories().stream().map(x -> x.getId()).collect(Collectors.toList());
-            List<Category> categories = categoryRepository.findAllById(categoryIds);
-            existingProduct.setCategories(categories);
+        if(categories != null) {
+            List<Category> categoryNews = categoryRepository.findAllById(categories);
+            existingProduct.setCategories(categoryNews);
         }
 
 
         // Xoá ảnh và thêm mới
-        imageRepository.deleteAllByProduct(existingProduct);
-        if (updatedProduct.getImages() != null) {
-            for (Images img : updatedProduct.getImages()) {
-                img.setProduct(existingProduct);
+        if(images != null) {
+            imageRepository.deleteAllByProduct(existingProduct);
+            List<Images> imagesNew = new ArrayList<>();
+            for (MultipartFile image : images) {
+                try{
+                    Images img =new Images();
+                    img.setBaseImage(image.getBytes());
+                    img.setProduct(existingProduct);
+                    imagesNew.add(img);
+                }catch (IOException e){
+                    throw new RuntimeException("loi up anh",e);
+                }
             }
-            existingProduct.setImages(updatedProduct.getImages());
+            existingProduct.setImages(imagesNew);
         }
 
-        // Xoá variant cũ và thêm mới
-        productVariantRepository.deleteAllByProduct(existingProduct);
-        if (updatedProduct.getVariants() != null) {
-            for (ProductVariant variant : updatedProduct.getVariants()) {
-                variant.setProduct(existingProduct);
-            }
-            existingProduct.setVariants(updatedProduct.getVariants());
-        }
+
+
 
         // Tính lại tổng quantity
-        long totalQuantity = updatedProduct.getVariants() != null
-                ? updatedProduct.getVariants().stream().mapToLong(ProductVariant::getStockQuantity).sum()
+        long totalQuantity = existingProduct.getVariants() != null
+                ? existingProduct.getVariants().stream().mapToLong(ProductVariant::getStockQuantity).sum()
                 : 0;
         existingProduct.setQuantity(totalQuantity);
 
@@ -120,14 +193,51 @@ public class ProductService {
         meta.setTotal(products.getTotalElements());
         meta.setPages(products.getTotalPages());
 
+        if(pageable.getSort().isSorted()){
+            meta.setSort(pageable.getSort().toString());
+        }
         resultPaginationDTO.setMeta(meta);
-        resultPaginationDTO.setResult(products.getContent());
+
+        List<ProductResponse> productResponses = products.getContent().stream().map(this::convertProductToProductResponse).collect(Collectors.toList());
+        resultPaginationDTO.setResult(productResponses);
+        return resultPaginationDTO;
+    }
+    public ResultPaginationDTO getBestSellingProducts(String name, Double minPrice, Double maxPrice, Long categoryId,int page,int size,String sort) {
+        Pageable pageable;
+        if (sort != null && !sort.isEmpty()) {
+            String[] parts = sort.split(",");
+            String sortField = parts[0];
+            Sort.Direction direction = parts.length > 1 && parts[1].equalsIgnoreCase("desc")
+                    ? Sort.Direction.DESC
+                    : Sort.Direction.ASC;
+
+            pageable = PageRequest.of(page - 1, size, Sort.by(direction, sortField));
+        } else {
+            pageable = PageRequest.of(page - 1, size);
+        }
+
+        Page<Product> products = productRepository.findBestSellingProduct(
+                categoryId, name, minPrice, maxPrice, pageable);
+        ResultPaginationDTO resultPaginationDTO = new ResultPaginationDTO();
+        ResultPaginationDTO.Meta meta = new ResultPaginationDTO.Meta();
+        meta.setPage(pageable.getPageNumber());
+        meta.setPageSize(pageable.getPageSize());
+        meta.setTotal(products.getTotalElements());
+        meta.setPages(products.getTotalPages());
+
+        if(pageable.getSort().isSorted()){
+            meta.setSort(pageable.getSort().toString());
+        }
+        resultPaginationDTO.setMeta(meta);
+        List<ProductResponse> productResponses = products.getContent().stream().map(this::convertProductToProductResponse).collect(Collectors.toList());
+        resultPaginationDTO.setResult(productResponses);
         return resultPaginationDTO;
     }
 
-    public Product getProduct(Long id) throws AppException {
-        return this.productRepository.findById(id)
+    public ProductResponse getProduct(Long id) throws AppException {
+        Product product =  this.productRepository.findById(id)
                 .orElseThrow(() -> new AppException("Product không tồn tại với id = " + id));
+        return convertProductToProductResponse(product);
     }
     public void deleteProduct(Long id) throws AppException {
         Product product = this.productRepository.findById(id)
